+++ src/stages/analysis/types/typebuilder.c
@@ -94,7 +94,7 @@ static int typebuilder_cmp_fun(void* lhs, void* rhs)
   typedesc_fun_t* lhs_desc = (typedesc_fun_t*)lhs;
   typedesc_fun_t* rhs_desc = (typedesc_fun_t*)rhs;
 
-  size_t lhs_hash = hash_digest(&lhs_desc->return_type, sizeof(typedesc_t*));
+  size_t lhs_hash = (size_t)hash_digest(&lhs_desc->return_type, sizeof(typedesc_t*));
 
   VECTOR_FOR_LOOP(i, lhs_desc->param_types)
   {
@@ -105,7 +105,7 @@ static int typebuilder_cmp_fun(void* lhs, void* rhs)
   lhs_hash = hash_combine_with_data(lhs_hash, &lhs_desc->is_vararg, sizeof(bool));
   lhs_hash = hash_combine_with_data(lhs_hash, &lhs_desc->callconv, sizeof(callconv_kind_t));
 
-  size_t rhs_hash = hash_digest(&rhs_desc->return_type, sizeof(typedesc_t*));
+  size_t rhs_hash = (size_t)hash_digest(&rhs_desc->return_type, sizeof(typedesc_t*));
 
   VECTOR_FOR_LOOP(i, rhs_desc->param_types)
   {
@@ -130,7 +130,7 @@ static int typebuilder_cmp_struct(void* lhs, void* rhs)
   typedesc_struct_t* lhs_desc = (typedesc_struct_t*)lhs;
   typedesc_struct_t* rhs_desc = (typedesc_struct_t*)rhs;
 
-  size_t lhs_hash = hash_digest(&lhs_desc->node, sizeof(ast_node_t*));
+  size_t lhs_hash = (size_t)hash_digest(&lhs_desc->node, sizeof(ast_node_t*));
 
   VECTOR_FOR_LOOP(i, lhs_desc->field_types)
   {
@@ -138,7 +138,7 @@ static int typebuilder_cmp_struct(void* lhs, void* rhs)
     lhs_hash = hash_combine_with_data(lhs_hash, &field_desc, sizeof(typedesc_t*));
   }
 
-  size_t rhs_hash = hash_digest(&rhs_desc->node, sizeof(ast_node_t*));
+  size_t rhs_hash = (size_t)hash_digest(&rhs_desc->node, sizeof(ast_node_t*));
 
   VECTOR_FOR_LOOP(i, rhs_desc->field_types)
   {
@@ -160,7 +160,7 @@ static int typebuilder_cmp_union(void* lhs, void* rhs)
   typedesc_union_t* lhs_desc = (typedesc_union_t*)lhs;
   typedesc_union_t* rhs_desc = (typedesc_union_t*)rhs;
 
-  size_t lhs_hash = hash_digest(&lhs_desc->node, sizeof(ast_node_t*));
+  size_t lhs_hash = (size_t)hash_digest(&lhs_desc->node, sizeof(ast_node_t*));
 
   VECTOR_FOR_LOOP(i, lhs_desc->field_types)
   {
@@ -168,7 +168,7 @@ static int typebuilder_cmp_union(void* lhs, void* rhs)
     lhs_hash = hash_combine_with_data(lhs_hash, &field_desc, sizeof(typedesc_t*));
   }
 
-  size_t rhs_hash = hash_digest(&rhs_desc->node, sizeof(ast_node_t*));
+  size_t rhs_hash = (size_t)hash_digest(&rhs_desc->node, sizeof(ast_node_t*));
 
   VECTOR_FOR_LOOP(i, rhs_desc->field_types)
   {
@@ -190,8 +190,8 @@ static int typebuilder_cmp_enum(void* lhs, void* rhs)
   typedesc_enum_t* lhs_desc = (typedesc_enum_t*)lhs;
   typedesc_enum_t* rhs_desc = (typedesc_enum_t*)rhs;
 
-  size_t lhs_hash = hash_digest(&lhs_desc->node, sizeof(ast_node_t*));
-  size_t rhs_hash = hash_digest(&rhs_desc->node, sizeof(ast_node_t*));
+  size_t lhs_hash = (size_t)hash_digest(&lhs_desc->node, sizeof(ast_node_t*));
+  size_t rhs_hash = (size_t)hash_digest(&rhs_desc->node, sizeof(ast_node_t*));
 
   if (lhs_hash < rhs_hash)
     return -1;
@@ -605,7 +605,7 @@ typedesc_t* typebuilder_build_union(typebuilder_t* builder, ast_node_t* node, ty
 
   for (size_t i = 0; i < field_count; i++)
   {
-    size_t field_size = LLVMSizeOfTypeInBits(builder->llvm_layout, field_types[i]->llvm_type);
+    size_t field_size = (size_t)LLVMSizeOfTypeInBits(builder->llvm_layout, field_types[i]->llvm_type);
 
     if (field_size > max_field_size)
     {
